infra/requirements.txt
aws-cdk-lib>=2.236.0
aws-cdk.aws-lambda-python-alpha>=2.248.0a0
python-dotenv>=1.1.1
